modified scripts.js, parseInt changed to parseFloat

+++ js/scripts.js
@@ -1,7 +1,7 @@
 function triangle() {
-var a =parseInt(document.getElementById("a").value);
-var b =parseInt(document.getElementById("b").value);
-var c =parseInt(document.getElementById("c").value);
+var a =parseFloat(document.getElementById("a").value);
+var b =parseFloat(document.getElementById("b").value);
+var c =parseFloat(document.getElementById("c").value);
 
   if (a===b && b===c && c===b) {
     alert("equilateral triangle")
